+++ copy-deps-scripts/clone.sh
@@ -8,26 +8,28 @@ cd repos
 git clone https://github.com/clojure/tools.analyzer.git
 cd tools.analyzer
 #git checkout tools.analyzer-0.6.3
-# One more commit after 0.6.3
-git checkout 4df0489243254cbc5112e44907012dcf2d4bd613
+# Several commits after 0.6.3, up through Nov 15 2014
+git checkout fd763e6a06319679f933f7fb9bf29910a004b262
 cd ..
 
 git clone https://github.com/clojure/tools.analyzer.jvm.git
 cd tools.analyzer.jvm
 #git checkout tools.analyzer.jvm-0.6.4
-# A few commits after 0.6.4
-git checkout c242be587db8406a778dc8edac190edf53f8b697
+# Several commits after 0.6.4, up through Nov 15 2014
+git checkout 304b4726f4d80ba32705e0335d156b1dc5bdcbb8
 cd ..
 
 git clone https://github.com/clojure/core.memoize.git
 cd core.memoize
-git checkout core.memoize-0.5.6
-patch -p1 < ../../core.memoize-fix-reflection-warning.patch
+#git checkout core.memoize-0.5.6
+# Many commits after 0.5.6, up through Nov 16 2014
+git checkout c0a91d09a666490c4b2702f4f05f5430b5fb9e3d
 cd ..
 
 git clone https://github.com/clojure/core.cache.git
 cd core.cache
-git checkout core.cache-0.6.3
+# Latest as of Nov 17 2014 has only doc changes from core.cache-0.6.4
+git checkout core.cache-0.6.4
 cd ..
 
 git clone https://github.com/clojure/data.priority-map.git